tests: Build Mariner rootfs
In summary, this commit does the following:

 * Modifies `versions.yaml` so that the target image/initrd now also
   depends on a specified host OS, not just architecture.
 * Sets the custom initrd path via annotation in `setup.sh`.
 * Adds a new `rootfs-initrd-mariner` build target.
 * Refactors the image builder so that its caller can specifiy the
   host OS, instead of fetching it from `versions.yaml` itself.

Currently, the Mariner Kata host only supports Mariner guests. Hence, I
modified `versions.yaml` so that the target image/initrd would now also
depend on a specified host OS, not just the architecture. Given that the
binaries seems to run well on Mariner, this avoids having to set up a
self-hosted Github runner.

At test time, `setup.sh` dynamically sets the image path based on
`versions.yaml` using a k8s annotation. This keeps both the workflow
YAML and the test scripts cleaner.

For the build, a new `rootfs-initrd-mariner` target has been created.
This carries the new host OS variable and propagates it down to the
Docker builder and then the initrd builder.

Signed-off-by: Aurélien Bombo <abombo@microsoft.com>